moved AWSSessionCredentialsFactory under aws package
