+++ digdag-standards/src/main/java/io/digdag/standards/operator/redshift/BaseRedshiftLoadOperator.java
@@ -14,8 +14,8 @@ import io.digdag.spi.SecretProvider;
 import io.digdag.spi.TaskExecutionException;
 import io.digdag.spi.TaskResult;
 import io.digdag.spi.TemplateEngine;
-import io.digdag.standards.operator.AWSSessionCredentialsFactory;
-import io.digdag.standards.operator.AWSSessionCredentialsFactory.AcceptableUri;
+import io.digdag.standards.operator.aws.AWSSessionCredentialsFactory;
+import io.digdag.standards.operator.aws.AWSSessionCredentialsFactory.AcceptableUri;
 import io.digdag.standards.operator.jdbc.AbstractJdbcJobOperator;
 import io.digdag.standards.operator.jdbc.DatabaseException;
 import io.digdag.standards.operator.jdbc.LockConflictException;
