+++ digdag-standards/src/main/java/io/digdag/standards/operator/redshift/RedshiftLoadOperatorFactory.java
@@ -18,8 +18,8 @@ import io.digdag.spi.OperatorFactory;
 import io.digdag.spi.SecretProvider;
 import io.digdag.spi.TaskExecutionException;
 import io.digdag.spi.TemplateEngine;
-import io.digdag.standards.operator.AWSSessionCredentialsFactory;
-import io.digdag.standards.operator.AWSSessionCredentialsFactory.AcceptableUri;
+import io.digdag.standards.operator.aws.AWSSessionCredentialsFactory;
+import io.digdag.standards.operator.aws.AWSSessionCredentialsFactory.AcceptableUri;
 import io.digdag.util.RetryExecutor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
